requirements.txt
fastapi
uvicorn[standard]
python-dotenv
spacetrack
skyfield
pydantic

